requirements.txt
numpy
pypdf2
pymupdf
Pillow
pikepdf
tqdm
opencv-python
reportlab
pyyaml
python-docx
pandas
